requirements.txt
streamlit==1.35.0
tensorflow-cpu==2.16.1
numpy==1.26.4
Pillow==10.3.0
